VFSGLOBAL/requirements.txt
aiogram==2.13
selenium
requests
python-dotenv